# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TownRide Carpooling API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Get my profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateMyProfile
      tags: [users]
      summary: Update my profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{userId}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips:
    get:
      operationId: searchTrips
      tags: [trips]
      summary: Search trips
      parameters:
        - name: origin
          in: query
          schema:
            type: string
        - name: destination
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripListResponse"

    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripBody"
      responses:
        "201":
          description: Trip created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/{tripId}:
    get:
      operationId: getTripById
      tags: [trips]
      summary: Get trip by ID
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateTripStatus
      tags: [trips]
      summary: Update trip status
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTripStatusBody"
      responses:
        "200":
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Delete a trip (admin)
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/my/driver:
    get:
      operationId: getMyDriverTrips
      tags: [trips]
      summary: Get my trips as driver
      responses:
        "200":
          description: My driver trips
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        replitUserId:
          type: string
        username:
          type: string
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        isDriver:
          type: boolean
        isAdmin:
          type: boolean
        vehicleModel:
          type: string
          nullable: true
        vehicleColor:
          type: string
          nullable: true
        licensePlate:
          type: string
          nullable: true
        totalSeats:
          type: integer
          nullable: true
        acceptsPackages:
          type: boolean
        avgRating:
          type: number
          nullable: true
        totalRatings:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - replitUserId
        - username
        - isDriver
        - isAdmin
        - acceptsPackages
        - totalRatings
        - createdAt
    UpdateProfileBody:
      type: object
      properties:
        isDriver:
          type: boolean
        vehicleModel:
          type: string
        vehicleColor:
          type: string
        licensePlate:
          type: string
        totalSeats:
          type: integer
        acceptsPackages:
          type: boolean
    Trip:
      type: object
      properties:
        id:
          type: integer
        driverId:
          type: integer
        origin:
          type: string
        destination:
          type: string
        date:
          type: string
        time:
          type: string
        availableSeats:
          type: integer
        totalSeats:
          type: integer
        pricePerSeat:
          type: number
        meetingPoint:
          type: string
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
        acceptsPackages:
          type: boolean
        driverName:
          type: string
          nullable: true
        driverAvatarUrl:
          type: string
          nullable: true
        driverRating:
          type: number
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - driverId
        - origin
        - destination
        - date
        - time
        - availableSeats
        - totalSeats
        - pricePerSeat
        - meetingPoint
        - status
        - acceptsPackages
        - createdAt
    TripDetail:
      allOf:
        - $ref: "#/components/schemas/Trip"
        - type: object
          properties:
            bookings:
              type: array
              items:
                $ref: "#/components/schemas/Booking"
    TripListResponse:
      type: object
      properties:
        trips:
          type: array
          items:
            $ref: "#/components/schemas/Trip"
      required: [trips]
    CreateTripBody:
      type: object
      properties:
        origin:
          type: string
        destination:
          type: string
        date:
          type: string
        time:
          type: string
        availableSeats:
          type: integer
        pricePerSeat:
          type: number
        meetingPoint:
          type: string
        acceptsPackages:
          type: boolean
      required:
        - origin
        - destination
        - date
        - time
        - availableSeats
        - pricePerSeat
        - meetingPoint
    UpdateTripStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
      required: [status]
    Booking:
      type: object
      properties:
        id:
          type: integer
        tripId:
          type: integer
        passengerId:
          type: integer
        status:
          type: string
          enum: [confirmed, cancelled]
        seatsBooked:
          type: integer
        passengerName:
          type: string
          nullable: true
        passengerAvatarUrl:
          type: string
          nullable: true
        trip:
          $ref: "#/components/schemas/Trip"
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - tripId
        - passengerId
        - status
        - seatsBooked
        - createdAt
...